Break out the critical point finding from `homotopy.jl`
